chore: create integration test helper file with setup and teardown
